+++ dandy/agent/agent.py
@@ -1,5 +1,0 @@
-from dandy.core.processor.processor import BaseProcessor
-
-
-class BaseAgent(BaseProcessor):
-    pass