+++ dandy/constants.py
@@ -1,5 +1,7 @@
 __VERSION__ = '0.13.3'
 
+from dandy.recorder.recorder import Recorder
+
 # Cache
 
 SQLITE_CACHE_TABLE_NAME = 'cache'
@@ -17,8 +19,8 @@ CLI_DEFAULT_ENV_FILE_NAMES = [
 
 # Debug
 
-DEBUG_OUTPUT_DIRECTORY = '.dandy_debug_output'
-DEBUG_DEFAULT_NAME = 'default'
+RECORDER_OUTPUT_DIRECTORY = '.dandy_debug_output'
+RECORDING_DEFAULT_NAME = 'default'
 
 # LLM