+++ dandy/debug/exceptions.py
@@ -1,9 +1,0 @@
-from dandy.core.exceptions import DandyCriticalException, DandyRecoverableException
-
-
-class DebugCriticalException(DandyCriticalException):
-    pass
-
-
-class DebugRecoverableException(DandyRecoverableException):
-    pass