+++ dandy/recorder/exceptions.py
@@ -1,0 +1,9 @@
+from dandy.core.exceptions import DandyCriticalException, DandyRecoverableException
+
+
+class RecorderCriticalException(DandyCriticalException):
+    pass
+
+
+class RecorderRecoverableException(DandyRecoverableException):
+    pass